dom/base/test/test_bug982153.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=982153
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 982153</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script>

var sc = document.createElement("script");
var error = null;
sc.textContent = "try {\n reference_error; } catch(e) { error = e; }";
document.documentElement.appendChild(sc);
is(error.lineNumber, 2, "Error line number must be correct");

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=982153">Mozilla Bug 982153</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
</body>
</html>
